refactor(remediationorchestrator): extract notification_creation.go from reconciler.go
Phase 3 file split: moves hasNotificationRef, ensureNotificationsCreated,
buildNotificationRef, buildTimeoutContext, and createPhaseTimeoutNotification
into their own file. Pure structural move: 376 Ginkgo specs still pass.
